backend/pyproject.toml
[tool.poetry]
name = "goodstewards-backend"
version = "0.1.0"
description = "Backend for GoodStewards application"
authors = ["Gemini CLI Agent <gemini-cli@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlmodel = "^0.0.18"
psycopg = {extras = ["binary", "pool"], version = "^3.1.18"}
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
baml-py = "^0.202.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
boto3 = "^1.34.108"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
black = "^24.0.0"
ruff = "^0.3.0"
mypy = "^1.8.0"
bandit = "^1.7.5"
safety = "^2.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
